requirements.txt
r2pipe==1.8.0
requests==2.28.1
vt-py==0.17.5
yara-python==4.2.0
